+++ src/main/java/se/iths/weblab2/FishDto.java
@@ -5,9 +5,9 @@ public class FishDto {
     private String name;
     private String type;
     private String gender;
-    private float weight;
+    private double weight;
 
-    public FishDto(int id, String name, String type, String gender, float weight) {
+    public FishDto(int id, String name, String type, String gender, double weight) {
         this.id = id;
         this.name = name;
         this.type = type;
@@ -47,11 +47,11 @@ public class FishDto {
         this.gender = gender;
     }
 
-    public float getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 }
